+++ craco.config.js
@@ -3,11 +3,12 @@ const path = require("path");
 module.exports = {
   webpack: {
     alias: {
-      "@components": path.resolve(__dirname, "src/components"),
-      "@pages": path.resolve(__dirname, "src/pages"),
       "@auth": path.resolve(__dirname, "src/auth/index"),
-      "@models": path.resolve(__dirname, "src/models"),
+      "@components": path.resolve(__dirname, "src/components"),
       "@data": path.resolve(__dirname, "src/data"),
+      "@db": path.resolve(__dirname, "src/db"),
+      "@models": path.resolve(__dirname, "src/models"),
+      "@pages": path.resolve(__dirname, "src/pages"),
       "@stores": path.resolve(__dirname, "src/stores"),
     },
   },