+++ src/models/word.ts
@@ -1,4 +1,4 @@
-import { createHash } from "crypto";
+import hashjs from "hash.js";
 
 export interface Word {
   word: string;
@@ -10,6 +10,29 @@ export interface Word {
 
 export namespace Word {
   export function hash(word: string): string {
-    return createHash("sha256").update(word).digest("hex");
+    return hashjs.sha256().update(word).digest("hex");
+  }
+
+  export function New(word: string): Word {
+    return {
+      word: word,
+      highestWpm: 0,
+      highestWpmDatetime: new Date(0),
+      lastPracticeWpm: 0,
+      lastPracticeDatetime: new Date(0),
+    };
+  }
+
+  export function updateWpm(word: Word, wpm: number): Word {
+    const date = new Date();
+    if (wpm > word.highestWpm) {
+      word.highestWpm = wpm;
+      word.highestWpmDatetime = date;
+    }
+
+    word.lastPracticeWpm = wpm;
+    word.lastPracticeDatetime = date;
+
+    return word;
   }
 }